refactor: merge all account *Overview components
