--- ui/components/multichain/account-overview/account-overview-eth.stories.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { AccountOverviewEth } from './account-overview-eth'
-import { AccountOverviewCommonProps } from './common';
-
-export default {
-  title: 'Components/Multichain/AccountOverviewEth',
-  component: AccountOverviewEth,
-};
-
-export const DefaultStory = (
-  args: JSX.IntrinsicAttributes & AccountOverviewCommonProps
-) => <AccountOverviewEth {...args} />;
