--- ui/components/multichain/account-overview/account-overview-non-evm.stories.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { AccountOverviewNonEvm } from './account-overview-non-evm';
-import { AccountOverviewCommonProps } from './common';
-import { BtcAccountType, SolAccountType, TrxAccountType } from '@metamask/keyring-api';
-
-export default {
-  title: 'Components/Multichain/AccountOverviewNonEvm',
-  component: AccountOverviewNonEvm,
-  args: {
-    accountType: BtcAccountType.P2wpkh,
-  },
-};
-
-export const DefaultStory = (
-  args: JSX.IntrinsicAttributes &
-    AccountOverviewCommonProps & {
-      accountType: BtcAccountType.P2wpkh | SolAccountType.DataAccount | TrxAccountType.Eoa;
-    },
-) => <AccountOverviewNonEvm {...args} />;
