+++ ui/components/multichain/account-overview/account-overview-tabs.tsx
@@ -156,8 +156,8 @@ export const AccountOverviewTabs = ({
           >
             <Box marginBottom={2}>
               <AssetList
-                showTokensLinks={showTokensLinks ?? true}
                 onClickAsset={onClickAsset}
+                showTokensLinks={showTokensLinks}
                 safeChains={safeChains}
               />
             </Box>
@@ -182,8 +182,8 @@ export const AccountOverviewTabs = ({
           >
             <Box>
               <DeFiTab
-                showTokensLinks={showTokensLinks ?? true}
                 onClickAsset={onClickDeFi}
+                showTokensLinks={showTokensLinks}
                 safeChains={safeChains}
               />
             </Box>
