--- ui/components/multichain/account-overview/account-overview-unknown.stories.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import { AccountOverviewUnknown } from './account-overview-unknown';
-import { AccountOverviewCommonProps } from './common';
-
-export default {
-  title: 'Components/Multichain/AccountOverviewUnknown',
-  component: AccountOverviewUnknown,
-  parameters: {
-    initialEntries: ['/'],
-    path: '*',
-  },
-};
-
-export const DefaultStory = (
-  args: JSX.IntrinsicAttributes & AccountOverviewCommonProps,
-) => <AccountOverviewUnknown {...args} />;
